requirements.txt
Flask==2.1.2
flask-cors==3.1.1
pycryptodome==3.10.1
gunicorn